Problema al calcular la cantidad de dias entre fechas fixeado

+++ src/main/java/com/example/hoteleria/entities/Reserva.java
@@ -39,7 +39,7 @@ public class Reserva {
         habitacion.getReservas().remove(this);
     }
 
-    public void calcularPrecioTotal(int cantidadDeDias){
+    public void calcularPrecioTotal(long cantidadDeDias){
         BigDecimal total = new BigDecimal(0.00);
         for(Habitacion habitacion : this.habitaciones){
             total = total.add(habitacion.getPrecio());
